requirements.txt
Flask==2.2.5
numpy==1.26.4
python-dotenv==1.0.1
